fix: update pdf also during version sync
