+++ package.json
@@ -21,7 +21,7 @@
     "update-version": "node ./utils/update-version.js",
     "release": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it",
     "release:dry-run": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it --dry-run",
-    "generate-pdf": "node ./utils/generate-pdf"
+    "generate-pdf": "node ./utils/generate-pdf-cli"
   },
   "devDependencies": {
     "@markedjs/html-differ": "^3.0.4",
